requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
black==22.1.0
click==8.0.4; python_version >= '3.6'
dj-database-url==0.5.0
dj-static==0.0.6
django-cors-headers==3.11.0
django-dotenv==1.4.2
django==3.2
djangorestframework==3.13.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
pillow==9.0.1
platformdirs==2.5.1; python_version >= '3.7'
psycopg2==2.9.3
pytz==2022.1
setuptools==61.0.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
static3==0.7.0
tomli==2.0.1; python_version >= '3.7'
